move gelu defition to .h
